Sanitize string ctx input
I noticed a part of the guide that mentioned a security issue stemming
from string input sanitizing for the `ctx` argument passed to `job`.
Link: https://github.com/wilk/microjob/issues/2

I was surprised to see that when assigning the string context variable
that the only thing that seemed to go on was wrapping the string in a
set of single-quotes:
```js
case 'string':
  variable = `'${config.ctx[key]}'`
  break
```
which allows the exploit shown in #2 to occur. This PR simply uses
`JSON.stringify`, just like is already done with objects to stop anyone
from doing any quote-related funny business.

This commit passes the test added in the previous commit.